Simplify configuration method. Split config into its own test. Add before setup to tests
