+++ Rakefile
@@ -4,7 +4,7 @@ require 'rake/testtask'
 Rake::TestTask.new(:test) do |t|
   t.libs << "test"
   t.libs << "lib"
-  t.test_files = FileList['test/test_pirata.rb']
+  t.test_files = FileList['test/test_pirata.rb', 'test/test_pirata_config.rb']
 end
 
 task :default => :test
